src-tauri/resources/py_app/requirements.txt
huggingface_hub>=0.22.0
llama-cpp-python>=0.3.0

